Otimização da Trilha de Financeiro (Realizar Pagamento)

+++ src/pages/Finance/PerformPayment/index.jsx
@@ -2,34 +2,39 @@ import './performPayment.css'
 import ItemsPayment from '../../../components/ItemsPayment';
 import { useNavigate } from 'react-router-dom'
 import { useState } from 'react';
+import Swal from 'sweetalert2';
+import withReactContent from 'sweetalert2-react-content';
 
 const PerformPayment = () => {
+    
     const [selectedSubjects, setSelectedSubjects] = useState([]);
     const navigate = useNavigate();
+
+    const MySwal = withReactContent(Swal);
 
     const list = [
         {
             id: 1,
             name: 'Mensalidade Jan/234',
-            valor: '79.90',
+            valor: '79,90',
             status: 'Acordo'
         },
         {
             id: 2,
             name: 'Mensalidade Jan/235',
-            valor: '139.90',
+            valor: '139,90',
             status: 'Serviço'
         },
         {
             id: 3,
             name: 'Mensalidade Jan/236',
-            valor: '138.90',
+            valor: '138,90',
             status: 'Mensalidade'
         },
         {
             id: 4,
             name: 'Mensalidade Jan/237',
-            valor: '155.90',
+            valor: '155,90',
             status: 'Outros'
         }
     ];
@@ -44,24 +49,31 @@ const PerformPayment = () => {
         });
     };
 
-    const total = list.reduce((acc, curr) => {
-        if (selectedSubjects.includes(curr.id)) {
-            return acc + parseFloat(curr.valor.replace(',', '.'));
-        }
-        return acc;
+    const total = selectedSubjects.reduce((acc, id) => {
+        const item = list.find(item => item.id === id);
+        return acc + parseFloat(item.valor.replace(',', '.'));
     }, 0);
 
+    const formatValue = (value) => {
+        return value.toFixed(2).replace('.', ',');
+    };
+
     const handleNext = () => {
-        const selectedItems = list.filter(item => selectedSubjects.includes(item.id));
-        navigate('/financeiro/realizar-pagamento/detalhes-pagamento', {
-            state: { selectedItems }
-        });
-
-        const formatValue = (value) => {
-            return value.toFixed(2).replace('.', ',');
-          };
-
-        localStorage.setItem("total", formatValue(total));
+        if (selectedSubjects.length === 0) {
+            MySwal.fire({
+                icon: 'info',
+                title: 'Erro',
+                text: 'Você não selecionou nada',
+                confirmButtonText: 'OK'
+            });
+        } else {
+            const selectedItems = list.filter(item => selectedSubjects.includes(item.id));
+            localStorage.setItem("selectedItems", JSON.stringify(selectedItems));
+    
+            localStorage.setItem("total", formatValue(total));
+    
+            navigate('/financeiro/realizar-pagamento/detalhes-pagamento');
+        }
     };
 
     return (
@@ -79,7 +91,7 @@ const PerformPayment = () => {
                 <div className='footer-payment'>
                     <div className='total'>
                         <span>Total:</span>
-                        <span>R$ {total.toFixed(2)}</span>
+                        <span>R$ {formatValue(total)}</span>
                     </div>
                     <button onClick={handleNext} className='title-footer' > Próximo </button>
                 </div>
